delete offchain changesets
