+++ deployment/v2_0_0/changesets/apply_config_cl_mode_test.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 	"sync"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -280,230 +279,6 @@ func TestApplyVerifierConfig_PassesWhenNonTargetNOPMissingChainSupport(t *testin
 	assert.NotNil(t, output.DataStore)
 }
 
-func TestApplyExecutorConfig_FailsWhenNOPMissingChainSupport(t *testing.T) {
-	sel1 := chainsel.TEST_90000001.Selector
-	sel2 := chainsel.TEST_90000002.Selector
-
-	executorAdapter := &mockExecutorConfigAdapter{
-		deployedChains: map[string][]uint64{
-			"pool1": {sel1, sel2},
-		},
-		chainConfigs: map[uint64]adapters.ExecutorChainConfig{
-			sel1: {
-				OffRampAddress:       "0xOffRamp",
-				RmnAddress:           "0xRmn",
-				ExecutorProxyAddress: "0xExecutorProxy",
-			},
-			sel2: {
-				OffRampAddress:       "0xOffRamp",
-				RmnAddress:           "0xRmn",
-				ExecutorProxyAddress: "0xExecutorProxy",
-			},
-		},
-	}
-
-	registry := adapters.NewExecutorConfigRegistry()
-	registry.Register(chainsel.FamilyEVM, executorAdapter)
-
-	sel1Str := fmt.Sprintf("%d", sel1)
-	sel2Str := fmt.Sprintf("%d", sel2)
-	topo := newTopologyWithChainConfigs(
-		[]string{"nop1", "nop2"},
-		"pool1",
-		shared.NOPModeCL,
-		map[string]offchain.ChainExecutorPoolConfig{
-			sel1Str: {NOPAliases: []string{"nop1", "nop2"}, ExecutionInterval: 15 * time.Second},
-			sel2Str: {NOPAliases: []string{"nop1", "nop2"}, ExecutionInterval: 15 * time.Second},
-		},
-	)
-	env := newTestExecutorEnv(t, []uint64{sel1, sel2})
-
-	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{
-			Nodes: []*nodev1.Node{
-				{Id: "node-1", Name: "nop1"},
-				{Id: "node-2", Name: "nop2"},
-			},
-		}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{
-			ChainConfigs: []*nodev1.ChainConfig{
-				chainConfig("node-1", "90000001"),
-				chainConfig("node-2", "90000001"),
-				chainConfig("node-2", "90000002"),
-			},
-		}, nil,
-	)
-	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1", "node-2"}
-
-	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
-	_, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
-		Topology:          topo,
-		ExecutorQualifier: "pool1",
-	})
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "chain support validation failed")
-	assert.Contains(t, err.Error(), "nop1")
-}
-
-func TestApplyExecutorConfig_PassesWhenNOPSupportsExtraChains(t *testing.T) {
-	sel1 := chainsel.TEST_90000001.Selector
-	sel2 := chainsel.TEST_90000002.Selector
-
-	executorAdapter := &mockExecutorConfigAdapter{
-		deployedChains: map[string][]uint64{
-			"pool1": {sel1, sel2},
-		},
-		chainConfigs: map[uint64]adapters.ExecutorChainConfig{
-			sel1: {
-				OffRampAddress:       "0xOffRamp",
-				RmnAddress:           "0xRmn",
-				ExecutorProxyAddress: "0xExecutorProxy",
-			},
-			sel2: {
-				OffRampAddress:       "0xOffRamp",
-				RmnAddress:           "0xRmn",
-				ExecutorProxyAddress: "0xExecutorProxy",
-			},
-		},
-	}
-
-	registry := adapters.NewExecutorConfigRegistry()
-	registry.Register(chainsel.FamilyEVM, executorAdapter)
-
-	topo := newMinimalTopology([]string{"nop1", "nop2"}, "pool1", shared.NOPModeCL)
-	env := newTestExecutorEnv(t, []uint64{sel1, sel2})
-
-	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{
-			Nodes: []*nodev1.Node{
-				{Id: "node-1", Name: "nop1"},
-				{Id: "node-2", Name: "nop2"},
-			},
-		}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{
-			ChainConfigs: []*nodev1.ChainConfig{
-				chainConfig("node-1", "90000001"),
-				chainConfig("node-1", "90000002"),
-				chainConfig("node-1", "90000003"),
-				chainConfig("node-2", "90000001"),
-				chainConfig("node-2", "90000002"),
-			},
-		}, nil,
-	)
-	mockJD.EXPECT().ProposeJob(mock.Anything, mock.Anything).Return(
-		&jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "job-1"}}, nil,
-	)
-	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1", "node-2"}
-
-	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
-	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
-		Topology:          topo,
-		ExecutorQualifier: "pool1",
-	})
-	require.NoError(t, err)
-	assert.NotNil(t, output.DataStore)
-}
-
-func TestApplyExecutorConfig_SkipsChainSupportValidationForStandaloneNOPs(t *testing.T) {
-	sel1 := chainsel.TEST_90000001.Selector
-
-	executorAdapter := &mockExecutorConfigAdapter{
-		deployedChains: map[string][]uint64{
-			"pool1": {sel1},
-		},
-		chainConfigs: map[uint64]adapters.ExecutorChainConfig{
-			sel1: {
-				OffRampAddress:       "0xOffRamp",
-				RmnAddress:           "0xRmn",
-				ExecutorProxyAddress: "0xExecutorProxy",
-			},
-		},
-	}
-
-	registry := adapters.NewExecutorConfigRegistry()
-	registry.Register(chainsel.FamilyEVM, executorAdapter)
-
-	topo := newMinimalTopology([]string{"nop1", "nop2"}, "pool1", shared.NOPModeStandalone)
-	env := newTestExecutorEnv(t, []uint64{sel1})
-
-	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
-	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
-		Topology:          topo,
-		ExecutorQualifier: "pool1",
-	})
-	require.NoError(t, err)
-	assert.NotNil(t, output.DataStore)
-}
-
-func TestApplyExecutorConfig_PassesWhenNonTargetNOPMissingChainSupport(t *testing.T) {
-	sel1 := chainsel.TEST_90000001.Selector
-	sel2 := chainsel.TEST_90000002.Selector
-
-	executorAdapter := &mockExecutorConfigAdapter{
-		deployedChains: map[string][]uint64{
-			"pool1": {sel1, sel2},
-		},
-		chainConfigs: map[uint64]adapters.ExecutorChainConfig{
-			sel1: {
-				OffRampAddress:       "0xOffRamp",
-				RmnAddress:           "0xRmn",
-				ExecutorProxyAddress: "0xExecutorProxy",
-			},
-			sel2: {
-				OffRampAddress:       "0xOffRamp",
-				RmnAddress:           "0xRmn",
-				ExecutorProxyAddress: "0xExecutorProxy",
-			},
-		},
-	}
-
-	registry := adapters.NewExecutorConfigRegistry()
-	registry.Register(chainsel.FamilyEVM, executorAdapter)
-
-	topo := newMinimalTopology([]string{"nop1", "nop2"}, "pool1", shared.NOPModeCL)
-	env := newTestExecutorEnv(t, []uint64{sel1, sel2})
-
-	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{
-			Nodes: []*nodev1.Node{
-				{Id: "node-1", Name: "nop1"},
-				{Id: "node-2", Name: "nop2"},
-			},
-		}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{
-			ChainConfigs: []*nodev1.ChainConfig{
-				chainConfig("node-1", "90000001"),
-				chainConfig("node-1", "90000002"),
-			},
-		}, nil,
-	)
-	mockJD.EXPECT().ProposeJob(mock.Anything, mock.Anything).Return(
-		&jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "job-1"}}, nil,
-	)
-	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1", "node-2"}
-
-	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
-	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
-		Topology:          topo,
-		ExecutorQualifier: "pool1",
-		TargetNOPs:        []shared.NOPAlias{"nop1"},
-	})
-	require.NoError(t, err)
-	assert.NotNil(t, output.DataStore)
-}
-
 func TestApplyVerifierConfig_ProposesCorrectSpecToCorrectNode(t *testing.T) {
 	sel1 := chainsel.TEST_90000001.Selector
 
@@ -578,83 +353,6 @@ func TestApplyVerifierConfig_ProposesCorrectSpecToCorrectNode(t *testing.T) {
 	require.NotNil(t, nop2Req, "nop2 should have a proposal on node-2")
 	assert.Contains(t, nop2Req.Spec, "nop2-", "nop2's spec should reference nop2 in the job name")
 	assert.NotContains(t, nop2Req.Spec, "nop1-", "nop2's spec must not contain nop1's job name")
-}
-
-func TestApplyExecutorConfig_ProposesCorrectSpecToCorrectNode(t *testing.T) {
-	sel1 := chainsel.TEST_90000001.Selector
-
-	executorAdapter := &mockExecutorConfigAdapter{
-		deployedChains: map[string][]uint64{
-			"pool1": {sel1},
-		},
-		chainConfigs: map[uint64]adapters.ExecutorChainConfig{
-			sel1: {
-				OffRampAddress:       "0xOffRamp",
-				RmnAddress:           "0xRmn",
-				ExecutorProxyAddress: "0xExecutorProxy",
-			},
-		},
-	}
-
-	registry := adapters.NewExecutorConfigRegistry()
-	registry.Register(chainsel.FamilyEVM, executorAdapter)
-
-	topo := newMinimalTopology([]string{"nop1", "nop2"}, "pool1", shared.NOPModeCL)
-	env := newTestExecutorEnv(t, []uint64{sel1})
-
-	captured := &proposeJobCapture{}
-
-	mockJD := mocks.NewMockClient(t)
-	mockJD.EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{
-			Nodes: []*nodev1.Node{
-				{Id: "node-1", Name: "nop1"},
-				{Id: "node-2", Name: "nop2"},
-			},
-		}, nil,
-	)
-	mockJD.EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{
-			ChainConfigs: []*nodev1.ChainConfig{
-				chainConfig("node-1", "90000001"),
-				chainConfig("node-2", "90000001"),
-			},
-		}, nil,
-	)
-	mockJD.EXPECT().ProposeJob(mock.Anything, mock.Anything).RunAndReturn(
-		func(_ context.Context, req *jobpb.ProposeJobRequest, _ ...grpc.CallOption) (*jobpb.ProposeJobResponse, error) {
-			captured.capture(req)
-			return &jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "prop-" + req.NodeId}}, nil
-		},
-	)
-	env.Offchain = mockJD
-	env.NodeIDs = []string{"node-1", "node-2"}
-
-	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
-	_, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
-		Topology:          topo,
-		ExecutorQualifier: "pool1",
-	})
-	require.NoError(t, err)
-
-	requests := captured.get()
-	require.Len(t, requests, 2, "expected exactly 2 ProposeJob calls for 2 NOPs")
-
-	for _, req := range requests {
-		assert.Contains(t, []string{"node-1", "node-2"}, req.NodeId, "proposed to unexpected node")
-		assert.Contains(t, req.Spec, `type = "ccvexecutor"`, "spec should be an executor job")
-		assert.Contains(t, req.Spec, "0xOffRamp", "spec should contain offramp address")
-	}
-
-	nop1Req := captured.findByNodeID("node-1")
-	require.NotNil(t, nop1Req, "nop1 should have a proposal on node-1")
-	assert.Contains(t, nop1Req.Spec, "nop1-pool1-executor", "nop1's spec should have correct job name")
-	assert.NotContains(t, nop1Req.Spec, "nop2-pool1-executor", "nop1's spec must not contain nop2's job name")
-
-	nop2Req := captured.findByNodeID("node-2")
-	require.NotNil(t, nop2Req, "nop2 should have a proposal on node-2")
-	assert.Contains(t, nop2Req.Spec, "nop2-pool1-executor", "nop2's spec should have correct job name")
-	assert.NotContains(t, nop2Req.Spec, "nop1-pool1-executor", "nop2's spec must not contain nop1's job name")
 }
 
 func TestApplyVerifierConfig_ExtraNodeIDsDoNotReceiveProposals(t *testing.T) {
@@ -956,65 +654,6 @@ func TestApplyVerifierConfig_RevokeOrphanedJobsDoesNotAffectOtherCommittees(t *t
 	require.NotNil(t, nopJobs)
 	_, hasC2 := nopJobs[c2VerifierJob.JobID]
 	assert.True(t, hasC2, "other committee (c2) verifier job must remain after verifier c1 revoke orphaned")
-}
-
-func TestApplyExecutorConfig_RevokeOrphanedJobsDoesNotAffectVerifierJobs(t *testing.T) {
-	sel1 := chainsel.TEST_90000001.Selector
-	executorAdapter := &mockExecutorConfigAdapter{
-		deployedChains: map[string][]uint64{"pool1": {sel1}},
-		chainConfigs: map[uint64]adapters.ExecutorChainConfig{
-			sel1: {
-				OffRampAddress:       "0xOffRamp",
-				RmnAddress:           "0xRmn",
-				ExecutorProxyAddress: "0xExecutorProxy",
-			},
-		},
-	}
-	registry := adapters.NewExecutorConfigRegistry()
-	registry.Register(chainsel.FamilyEVM, executorAdapter)
-
-	ds := datastore.NewMemoryDataStore()
-	verifierJob := shared.JobInfo{
-		JobID:    shared.JobID("nop1-agg-c1-verifier"),
-		NOPAlias: shared.NOPAlias("nop1"),
-		Proposals: map[string]shared.ProposalRevision{
-			"p1": {Status: shared.JobProposalStatusApproved},
-		},
-	}
-	require.NoError(t, offchain.SaveJobs(ds, []shared.JobInfo{verifierJob}))
-
-	topo := newMinimalTopology([]string{"nop1"}, "pool1", shared.NOPModeCL)
-	env := newTestExecutorEnv(t, []uint64{sel1})
-	env.DataStore = ds.Seal()
-	env.Offchain = mocks.NewMockClient(t)
-	env.NodeIDs = []string{"node-1"}
-	env.Offchain.(*mocks.MockClient).EXPECT().ListNodes(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodesResponse{Nodes: []*nodev1.Node{{Id: "node-1", Name: "nop1"}}}, nil,
-	)
-	env.Offchain.(*mocks.MockClient).EXPECT().ListNodeChainConfigs(mock.Anything, mock.Anything).Return(
-		&nodev1.ListNodeChainConfigsResponse{
-			ChainConfigs: []*nodev1.ChainConfig{chainConfig("node-1", "90000001")},
-		}, nil,
-	)
-	env.Offchain.(*mocks.MockClient).EXPECT().ProposeJob(mock.Anything, mock.Anything).Return(
-		&jobpb.ProposeJobResponse{Proposal: &jobpb.Proposal{Id: "prop-1"}}, nil,
-	)
-
-	cs := changesets.ApplyExecutorConfig(registry, newTestChainFamilyRegistry())
-	output, err := cs.Apply(env, changesets.ApplyExecutorConfigInput{
-		Topology:           topo,
-		ExecutorQualifier:  "pool1",
-		RevokeOrphanedJobs: true,
-	})
-	require.NoError(t, err)
-	require.NotNil(t, output.DataStore)
-
-	allJobs, err := offchain.GetAllJobs(output.DataStore.Seal())
-	require.NoError(t, err)
-	nopJobs := allJobs[shared.NOPAlias("nop1")]
-	require.NotNil(t, nopJobs)
-	_, hasVerifier := nopJobs[verifierJob.JobID]
-	assert.True(t, hasVerifier, "verifier job must remain after executor revoke orphaned")
 }
 
 func TestApplyVerifierConfig_TargetNOPsSucceedsWhenNonTargetedNOPMissingFromJD(t *testing.T) {
