+++ deployment/v2_0_0/changesets/internal_functions_test.go
@@ -1,8 +1,8 @@
 package changesets
 
 // Internal tests for package-private helpers that cannot be reached from the
-// external _test package (convertTopologyMonitoring, mustDecodeHex,
-// signerFromJDIfMissing, fetchSigningKeysForNOPsByFamilies, deriveFamiliesFromSelectors).
+// external _test package (convertTopologyMonitoring, signerFromJDIfMissing,
+// fetchSigningKeysForNOPsByFamilies, deriveFamiliesFromSelectors).
 
 import (
 	"context"
@@ -42,17 +42,6 @@ func TestConvertTopologyMonitoring_NilReturnsEmpty(t *testing.T) {
 	result := convertTopologyMonitoring(nil)
 	assert.False(t, result.Enabled)
 	assert.Empty(t, result.Type)
-}
-
-// ---- mustDecodeHex ----
-
-func TestMustDecodeHex_ValidHex(t *testing.T) {
-	b := mustDecodeHex("deadbeef")
-	require.Equal(t, []byte{0xde, 0xad, 0xbe, 0xef}, b)
-}
-
-func TestMustDecodeHex_InvalidHexPanics(t *testing.T) {
-	require.Panics(t, func() { mustDecodeHex("ZZZZ") })
 }
 
 // ---- signerFromJDIfMissing ----
