+++ deployment/v2_0_0/offchain/env_metadata_util.go
@@ -11,122 +11,11 @@ import (
 )
 
 type OffchainConfigs struct {
-	Aggregators    map[string]*Committee                    `json:"aggregators,omitempty"`
-	Indexers       map[string]*IndexerGeneratedConfig       `json:"indexers,omitempty"`
-	TokenVerifiers map[string]*TokenVerifierGeneratedConfig `json:"tokenVerifiers,omitempty"`
-	NOPJobs        shared.NOPJobs                           `json:"nopJobs,omitempty"`
+	NOPJobs shared.NOPJobs `json:"nopJobs,omitempty"`
 }
 
 type CCVEnvMetadata struct {
 	OffchainConfigs *OffchainConfigs `json:"offchainConfigs,omitempty"`
-}
-
-func SaveAggregatorConfig(ds datastore.MutableDataStore, serviceIdentifier string, cfg *Committee) error {
-	ccvMeta, err := loadOrCreateCCVEnvMetadata(ds)
-	if err != nil {
-		return err
-	}
-
-	if ccvMeta.OffchainConfigs == nil {
-		ccvMeta.OffchainConfigs = &OffchainConfigs{}
-	}
-	if ccvMeta.OffchainConfigs.Aggregators == nil {
-		ccvMeta.OffchainConfigs.Aggregators = make(map[string]*Committee)
-	}
-
-	ccvMeta.OffchainConfigs.Aggregators[serviceIdentifier] = cfg
-
-	return persistCCVEnvMetadata(ds, ccvMeta)
-}
-
-func GetAggregatorConfig(ds datastore.DataStore, serviceIdentifier string) (*Committee, error) {
-	ccvMeta, err := loadCCVEnvMetadata(ds)
-	if err != nil {
-		return nil, err
-	}
-
-	if ccvMeta.OffchainConfigs == nil || ccvMeta.OffchainConfigs.Aggregators == nil {
-		return nil, fmt.Errorf("no aggregator configs found")
-	}
-
-	cfg, ok := ccvMeta.OffchainConfigs.Aggregators[serviceIdentifier]
-	if !ok {
-		return nil, fmt.Errorf("aggregator config %q not found", serviceIdentifier)
-	}
-
-	return cfg, nil
-}
-
-func SaveIndexerConfig(ds datastore.MutableDataStore, serviceIdentifier string, cfg *IndexerGeneratedConfig) error {
-	ccvMeta, err := loadOrCreateCCVEnvMetadata(ds)
-	if err != nil {
-		return err
-	}
-
-	if ccvMeta.OffchainConfigs == nil {
-		ccvMeta.OffchainConfigs = &OffchainConfigs{}
-	}
-	if ccvMeta.OffchainConfigs.Indexers == nil {
-		ccvMeta.OffchainConfigs.Indexers = make(map[string]*IndexerGeneratedConfig)
-	}
-
-	ccvMeta.OffchainConfigs.Indexers[serviceIdentifier] = cfg
-
-	return persistCCVEnvMetadata(ds, ccvMeta)
-}
-
-func GetIndexerConfig(ds datastore.DataStore, serviceIdentifier string) (*IndexerGeneratedConfig, error) {
-	ccvMeta, err := loadCCVEnvMetadata(ds)
-	if err != nil {
-		return nil, err
-	}
-
-	if ccvMeta.OffchainConfigs == nil || ccvMeta.OffchainConfigs.Indexers == nil {
-		return nil, fmt.Errorf("no indexer configs found")
-	}
-
-	cfg, ok := ccvMeta.OffchainConfigs.Indexers[serviceIdentifier]
-	if !ok {
-		return nil, fmt.Errorf("indexer config %q not found", serviceIdentifier)
-	}
-
-	return cfg, nil
-}
-
-func SaveTokenVerifierConfig(ds datastore.MutableDataStore, serviceIdentifier string, cfg *TokenVerifierGeneratedConfig) error {
-	ccvMeta, err := loadOrCreateCCVEnvMetadata(ds)
-	if err != nil {
-		return err
-	}
-
-	if ccvMeta.OffchainConfigs == nil {
-		ccvMeta.OffchainConfigs = &OffchainConfigs{}
-	}
-	if ccvMeta.OffchainConfigs.TokenVerifiers == nil {
-		ccvMeta.OffchainConfigs.TokenVerifiers = make(map[string]*TokenVerifierGeneratedConfig)
-	}
-
-	ccvMeta.OffchainConfigs.TokenVerifiers[serviceIdentifier] = cfg
-
-	return persistCCVEnvMetadata(ds, ccvMeta)
-}
-
-func GetTokenVerifierConfig(ds datastore.DataStore, serviceIdentifier string) (*TokenVerifierGeneratedConfig, error) {
-	ccvMeta, err := loadCCVEnvMetadata(ds)
-	if err != nil {
-		return nil, err
-	}
-
-	if ccvMeta.OffchainConfigs == nil || ccvMeta.OffchainConfigs.TokenVerifiers == nil {
-		return nil, fmt.Errorf("no token verifier configs found")
-	}
-
-	cfg, ok := ccvMeta.OffchainConfigs.TokenVerifiers[serviceIdentifier]
-	if !ok {
-		return nil, fmt.Errorf("token verifier config %q not found", serviceIdentifier)
-	}
-
-	return cfg, nil
 }
 
 func loadCCVEnvMetadata(ds datastore.DataStore) (*CCVEnvMetadata, error) {
@@ -167,10 +56,9 @@ func parseCCVEnvMetadata(metadata any) (*CCVEnvMetadata, error) {
 }
 
 // persistCCVEnvMetadata persists CCV metadata using shallow merge at the
-// offchainConfigs level. Known offchainConfigs keys (aggregators, indexers,
-// tokenVerifiers, nopJobs) are replaced fully — removing stale nested entries
-// when chains or jobs are removed. Unknown sibling keys under offchainConfigs
-// and non-CCV top-level keys are preserved.
+// offchainConfigs level. Known offchainConfigs keys (nopJobs) are replaced
+// fully — removing stale nested entries when jobs are removed. Unknown sibling
+// keys under offchainConfigs and non-CCV top-level keys are preserved.
 func persistCCVEnvMetadata(ds datastore.MutableDataStore, ccvMeta *CCVEnvMetadata) error {
 	var existingMeta map[string]any
 	envMeta, err := ds.EnvMetadata().Get()
@@ -198,27 +86,6 @@ func persistCCVEnvMetadata(ds datastore.MutableDataStore, ccvMeta *CCVEnvMetadat
 		}
 
 		oc := ccvMeta.OffchainConfigs
-		if oc.Aggregators != nil {
-			v, err := marshalToAny(oc.Aggregators)
-			if err != nil {
-				return fmt.Errorf("failed to convert aggregators: %w", err)
-			}
-			existingOC["aggregators"] = v
-		}
-		if oc.Indexers != nil {
-			v, err := marshalToAny(oc.Indexers)
-			if err != nil {
-				return fmt.Errorf("failed to convert indexers: %w", err)
-			}
-			existingOC["indexers"] = v
-		}
-		if oc.TokenVerifiers != nil {
-			v, err := marshalToAny(oc.TokenVerifiers)
-			if err != nil {
-				return fmt.Errorf("failed to convert token verifiers: %w", err)
-			}
-			existingOC["tokenVerifiers"] = v
-		}
 		if oc.NOPJobs != nil {
 			v, err := marshalToAny(oc.NOPJobs)
 			if err != nil {
