--- deployment/v2_0_0/offchain/indexer.go
@@ -1,10 +1,0 @@
-package offchain
-
-type IndexerVerifierConfig struct {
-	Name            string   `json:"name"`
-	IssuerAddresses []string `json:"issuerAddresses"`
-}
-
-type IndexerGeneratedConfig struct {
-	Verifiers []IndexerVerifierConfig `json:"verifiers"`
-}
